automation_testing_backend/test_automation/requirements.txt
pytest
requests
allure-pytest

# para instalar, usar pip install -r requirements.txt